Added package for calculating continued fraction representation of square roots and updated problem 64 to use big.Int

+++ problems/problem0064.go
@@ -23,7 +23,7 @@ func (p problem0064) Solve() string {
 
     count := 0
 
-    for i := 2; i <= limit; i++ {
+    for i := int64(2); i <= limit; i++ {
         rep := continuedfractions.ComputeSquareRootRepresenation(i)
 
         period := len(rep) - 1
